requirements.txt
numpy>=1.22
Pillow>=9.0
scikit-learn>=1.1
opencv-python>=4.5.0
scipy>=1.7.0

# Face detection (mediapipe is default, others are optional)
mediapipe>=0.10.0

# Deep feature extractors
torch>=2.0
torchvision>=0.15

# ONNX export
onnx>=1.12.0
onnxruntime>=1.12.0

# Experiment tracking
wandb>=0.16.0

# Optional dependencies (install manually if needed):
# facenet-pytorch>=2.5.0  # For MTCNN face detection
# retinaface-pytorch      # For RetinaFace (use pytorch version, not tensorflow)
# timm>=0.9.2             # For additional model backbones
